fix: base32 for ipfs:// urls and cidv1 everywhere
- make POST return a base32 encoded cidv1.
- default to using cidv1 everywhere.
- update to use the new ipfs-core module.

base36 is the right thing to use to respect the max authority length for ipns:// scheme urls.
The assumption is folks should use base32 for CIDs in ipfs:// scheme urls, and it's the default output for cidv1.

The new ipfs-core modules has the same api as the `ipfs` module but doesn't include the cli and http-server code with it.

At time of writing the tests pass apart from the last one, which hangs. I think it's due to an unrelated bug https://github.com/ipfs/js-ipfs/issues/3692

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>